Feature/native UI (#15)
* native UI using Piston for the raytracer

* 2D renderer
performance of serialization

* WIP 3D renderer

make all tests compile (sadly far away from passing)

* WIP 3D renderer - single threaded - not working
we have a piston window and a rotating monkey, but man thats slow and wrong and ugly all over the place

* WIP 3D renderer - lazy for Piston window is bad, now the fps from the piston game loop work

* WIP 3D renderer - fps kinda working

* native UI - but it's not working - what a surprise
flickering, some tiles are shown, then disappear

* native UI - and now its working, but why the flickering?

* native UI - add PNGs to gitignore

* native UI - fmt, different scene rendering

* clippy things

* native UI using Piston for the raytracer

* native UI - but it's not working - what a surprise
flickering, some tiles are shown, then disappear

* native UI - and now its working, but why the flickering?

* native UI - add PNGs to gitignore

* native UI - fmt, different scene rendering

* clippy things

* cleanup

* 2D renderer
performance of serialization

* WIP 3D renderer

make all tests compile (sadly far away from passing)

* WIP 3D renderer - single threaded - not working
we have a piston window and a rotating monkey, but man thats slow and wrong and ugly all over the place

* WIP 3D renderer - lazy for Piston window is bad, now the fps from the piston game loop work

* WIP 3D renderer - fps kinda working

* native UI - but it's not working - what a surprise
flickering, some tiles are shown, then disappear

* native UI - and now its working, but why the flickering?

* native UI - add PNGs to gitignore

* native UI - fmt, different scene rendering

* clippy things

* increase fps, more tiles

+++ micrograd-rs/src/micrograd_rs_v3.rs
@@ -5,7 +5,7 @@ use std::hash::{Hash, Hasher};
 use std::ops::{Add, AddAssign, Div, Mul, Neg, Sub, SubAssign};
 use std::rc::Rc;
 
-#[derive(PartialEq)]
+#[derive(PartialEq, Eq)]
 pub enum OpEnumV3 {
     NONE,
     ADD,
@@ -833,7 +833,7 @@ mod tests {
     // https://youtu.be/VMj-3S1tku0?t=1875
     #[test]
     pub fn test_video() {
-        let a = ValueRefV3::new_value(2.0 as f64, "a".to_string());
+        let a = ValueRefV3::new_value(2.0, "a".to_string());
         let b = ValueRefV3::new_value(-3.0, "b".to_string());
         let c = ValueRefV3::new_value(10.0, "c".to_string());
         let f = ValueRefV3::new_value(-2.0, "f".to_string());
@@ -852,7 +852,7 @@ mod tests {
 
     #[test]
     pub fn test_add() {
-        let a = ValueRefV3::new_value(2.0 as f64, "a".to_string());
+        let a = ValueRefV3::new_value(2.0, "a".to_string());
         let b = ValueRefV3::new_value(3.0, "b".to_string());
 
         let mut x = &a + &b;
@@ -863,7 +863,7 @@ mod tests {
 
     #[test]
     pub fn test_mul() {
-        let a = ValueRefV3::new_value(2.0 as f64, "a".to_string());
+        let a = ValueRefV3::new_value(2.0, "a".to_string());
         let b = ValueRefV3::new_value(3.0, "b".to_string());
 
         let mut x = &a * &b;
@@ -899,7 +899,7 @@ mod tests {
     #[test]
     pub fn test_value_plus_f64_rhs() {
         let a = ValueRefV3::new_value(3.0, "a".to_string());
-        let mut b = &a + 1.0 as f64;
+        let mut b = &a + 1.0;
         assert_two_float(b.borrow().data, 4.0);
         b.backward();
         draw_graph(b, "test_a_plus_f64_rhs".to_string());
@@ -908,7 +908,7 @@ mod tests {
     #[test]
     pub fn test_value_plus_f64_lhs() {
         let a = ValueRefV3::new_value(4.0, "a".to_string());
-        let mut b = 23.0 as f64 + &a;
+        let mut b = 23.0 + &a;
         b.backward();
         assert_two_float(b.borrow().data, 27.0);
 
@@ -918,7 +918,7 @@ mod tests {
     #[test]
     pub fn test_value_mul_f64_rhs() {
         let a = ValueRefV3::new_value(3.0, "a".to_string());
-        let mut b = &a * 3.0 as f64;
+        let mut b = &a * 3.0;
         assert_two_float(b.borrow().data, 9.0);
         b.backward();
         draw_graph(b, "test_a_mul_f64_rhs".to_string());
@@ -927,7 +927,7 @@ mod tests {
     #[test]
     pub fn test_value_mul_f64_lhs() {
         let a = ValueRefV3::new_value(4.0, "a".to_string());
-        let mut b = 23.0 as f64 * &a;
+        let mut b = 23.0 * &a;
         b.backward();
         assert_two_float(b.borrow().data, 92.0);
 
@@ -1011,8 +1011,8 @@ mod tests {
         d += 3.0 * &d + (&b - &a).relu(); //         d += 3 * d + (b - a).relu()
         let mut e = &c - &d; //         e = c - d
         let mut f = (&e).pow(2.0); //         f = e**2
-        let mut g = &f / 2.0 as f64; //         g = f / 2.0
-        g += 10.0 as f64 / &f; //         g += 10.0 / f
+        let mut g = &f / 2.0; //         g = f / 2.0
+        g += 10.0 / &f; //         g += 10.0 / f
 
         g.backward();
